feat(dice): generalize pick modifier
