+++ src/system/dice/plot-die.ts
@@ -1,6 +1,3 @@
-// Dialogs
-import { PickDiceResultDialog } from '@system/applications/dialogs/pick-dice-result';
-
 // Constants
 import { IMPORTED_RESOURCES } from '@system/constants';
 
@@ -13,10 +10,7 @@ const SIDES: Record<number, string> = {
     6: `<img src="${IMPORTED_RESOURCES.PLOT_DICE_OP_IN_CHAT}" />`,
 };
 
-const PICK_MODIFIER_REGEX = /(gm)?p(\d+)?/i;
-
-export interface PlotDieData
-    extends Partial<foundry.dice.terms.DiceTerm.TermData> {
+export interface PlotDieData extends Partial<foundry.dice.terms.Die.TermData> {
     /**
      * The number of dice of this term to roll
      * @default 1
@@ -29,7 +23,7 @@ export interface PlotDieData
     results?: foundry.dice.terms.DiceTerm.Result[];
 }
 
-export class PlotDie extends foundry.dice.terms.DiceTerm {
+export class PlotDie extends foundry.dice.terms.Die {
     public readonly isPlotDie = true;
 
     constructor(data: PlotDieData = {}) {
@@ -40,19 +34,6 @@ export class PlotDie extends foundry.dice.terms.DiceTerm {
     }
 
     static DENOMINATION = 'p';
-
-    static MODIFIERS = {
-        r: foundry.dice.terms.Die.prototype.reroll.bind(this),
-        rr: foundry.dice.terms.Die.prototype.rerollRecursive.bind(this),
-        k: foundry.dice.terms.Die.prototype.keep.bind(this),
-        kh: foundry.dice.terms.Die.prototype.keep.bind(this),
-        kl: foundry.dice.terms.Die.prototype.keep.bind(this),
-        d: foundry.dice.terms.Die.prototype.drop.bind(this),
-        dh: foundry.dice.terms.Die.prototype.drop.bind(this),
-        dl: foundry.dice.terms.Die.prototype.drop.bind(this),
-        p: 'pick',
-        gmp: 'pick',
-    };
 
     /* --- Accessors --- */
 
@@ -93,26 +74,5 @@ export class PlotDie extends foundry.dice.terms.DiceTerm {
         result: foundry.dice.terms.DiceTerm.Result,
     ): string {
         return SIDES[result.result];
-    }
-
-    /* --- Modifiers --- */
-
-    /**
-     * Shows a pop-up that allows the user to pick the result
-     */
-    async pick(modifier: string) {
-        const rgx = new RegExp(PICK_MODIFIER_REGEX);
-        const match = rgx.exec(modifier);
-        if (!match) return false;
-
-        const [gm, numStr] = match.slice(1);
-        const isGm = !!gm; // NOTE: Unused at this time
-        const amount = Math.min(parseInt(numStr) || 1, this.number ?? 0);
-
-        // Show dialog
-        await PickDiceResultDialog.show({
-            term: this,
-            amount,
-        });
     }
 }
